brIframe: Add resize on shader overlay
* Add resize callback on shader overlay
* Add stop animation tick when iframe loaded
* Separate shaders to utils file
